make admin, remove admin, remove user protected by admin

+++ index.js
@@ -28,7 +28,7 @@ function verifyJWT(req, res, next) {
     if (err) {
       return res.status(403).send({ message: 'Forbidden Access!!' })
     }
-    console.log(decoded) // bar
+    // console.log(decoded) // bar
     req.decoded = decoded;
     next();
   });
@@ -140,25 +140,59 @@ async function run() {
       const result = await usersCollection.find().toArray();
       res.send(result)
     })
+    // delete user
     app.delete('/user/delete/:email', verifyJWT, async (req, res) => {
       const email = req.params.email;
-      // console.log(email);
-      const filter = { email: email }
-      const result = await usersCollection.deleteOne(filter);
-      res.send(result)
+      const requesterEmail = req.decoded.email;
+      const requester = await usersCollection.findOne({email: requesterEmail});
+      if(requester.role !== 'admin'){
+        return res.status(403).send({message: 'notAllowed'})
+      }else{
+        const filter = { email: email }
+        const result = await usersCollection.deleteOne(filter);
+        res.send(result)
+      }
+    })
+
+    // check an existing user if he is admin 
+    app.get('/admin/:email', async(req, res) => {
+      const email = req.params.email;
+      const user = await usersCollection.findOne({email: email});
+      const isAdmin = user.role === 'admin';
+      res.send({admin: isAdmin})
     })
 
     // make user admin 
     app.put('/user/admin/:email', verifyJWT, async(req, res) => {
       const email = req.params.email;
-      // console.log(email);
-      const filter = {email: email}
-      const update = {$set: {role: 'admin'}}
-      const result = await usersCollection.updateOne(filter, update);
-      res.send(result);
+      // disallow normal user to make another admin 
+      const requester = req.decoded.email;
+      const requesterAccount = await usersCollection.findOne({email: requester});
+      if(requesterAccount.role === 'admin'){
+        const filter = {email: email}
+        const update = {$set: {role: 'admin'}}
+        const result = await usersCollection.updateOne(filter, update);
+        res.send(result);
+      }else{
+        res.status(403).send({message: 'notAllowed'})
+      }
     })
 
+    // remove as admin
+    app.put('/user/remove/:email', verifyJWT, async(req, res) => {
+      const email = req.params.email;
+      const requesterEmail = req.decoded.email;
+      const requester = await usersCollection.findOne({email: requesterEmail});
 
+      if(email === requesterEmail || requester.role !== 'admin'){
+        res.status(403).send({message: "You can't remove yourself"});
+      }else{
+        const filter = {email: email};
+        const update = {$unset: {role: 1}};
+        const result = await usersCollection.updateOne(filter, update);
+        res.send(result);
+      }
+    })
 
   } finally {
   }
